+++ xarrayutils/__init__.py
@@ -1,9 +1,9 @@
-__version__ = "0.1.0"
 from . import utils
 from . import numpy_utils
 from . import xmitgcm_utils
 from .utils import *
 
 from ._version import get_versions
-__version__ = get_versions()['version']
+
+__version__ = get_versions()["version"]
 del get_versions
